backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
boto3>=1.34.129
requests-oauthlib>=2.0.0
cryptography>=42.0.8
python-dotenv==1.0.0
pymongo==4.6.0
pydantic==2.5.0
email-validator>=2.2.0
pyjwt>=2.10.1
passlib>=1.7.4
tzdata>=2024.2
motor==3.3.2
pytest>=8.0.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
python-jose>=3.3.0
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
python-multipart==0.0.6
jq>=1.6.0
typer>=0.9.0
